requirements.txt
httpx
tortoise-orm
asyncpg
pydantic
pydantic-settings
loguru
aiofiles
aioschedule
beautifulsoup4
lxml
prometheus-client
python-dotenv
PyYAML
uvloop
motor
